// src/modules/user/enums/role.enum.ts
export enum RoleEnum {
  SUPER_USER = 'super_user',
  BUSINESS_OWNER = 'business_owner',
  SUPERVISOR = 'supervisor',
  PACKING_ADMIN = 'packing_admin',
  SHIPPING_ADMIN = 'shipping_admin',
  ADMIN = 'admin',
  MARKETING = 'marketing',
  EMPLOYEE = 'employee',
  HRD = 'hrd',
  CUSTOMER = 'customer',
}